🆓 startx: Do not use `exec` to stay in session if X fails
This is really useful if the X server fails.

For me, some PCs are okay with my `/etc/X11/xorg.conf`,
and some aren't (nvidia graphics differences probably, I don't know),
and just renaming the file to anything or bringing it back to normal
solves the issue.

To do that, though, I need to be able to actually log in.
When `exec` is used, you can't, and you're forced to use another user
which doesn't use X. For this case, `root` helped, but it's annoying to
switch and you might not have access to root.

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>